+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.10.2] - 2025-07-21
+
+### Fixed
+- (Sprite Editor) Fixed Set Raster commands clearing when duplicated
+- Proper asset stack texture handling
+- Various minor bugfixes
+
 ## [0.10.1] - 2025-04-01
 
 ### Fixed